Adding scripts and processes for Production builds
