+++ src/browser/main.js
@@ -1,49 +1,21 @@
-'use strict';
-
-var app = require('app');
-var yargs = require('yargs');
+var app  = require('app');
 var path = require('path');
+var fs   = require('fs');
 
-var shellStartTime = Date.now();
-
-process.on('uncaughtException', function (error) {
-  if (error == null) {
-    error = {};
-  }
-
-  if (error.message != null) {
-    console.error(error.message);
-  }
-
-  if (error.stack != null) {
-    console.error(error.stack);
-  }
-});
+// ROOT/src/browser/../../cache
+var cachePath = path.join(__dirname, '..', '..', 'cache');
+var devMode = (process.argv || []).indexOf('-r') !== -1;
 
 // Note: It's important that you don't do anything with Electron
 // unless it's after 'ready', or else mysterious bad things will happen
 // to you.
 app.on('ready', function () {
   // Enable ES6 from this point on
-  require('electron-compile').initWithOptions({
-    cacheDir: path.join(__dirname, '../../cache')
-  });
-  var Application = require('./application');
+  if (fs.statSyncNoException(cachePath) && !devMode) {
+    require('electron-compile').initForProduction(cachePath);
+  } else {
+    require('electron-compile').init();
+  }
 
-  var args = parseCommandLine();
-  new Application(args);
-
-  console.log('App load time: ' + (Date.now() - shellStartTime) + 'ms');
+  require('./app');
 });
-
-function parseCommandLine() {
-  var options = yargs(process.argv.slice(1)).wrap(100);
-
-  options.alias('t', 'test').boolean('t').describe('t', 'Run the specs and exit with error code on failures.');
-
-  var args = options.argv;
-
-  return {
-    test: args.test
-  };
-}
