+++ static/less/components/prompt-form.less
@@ -1,6 +1,4 @@
-@prompt-form-width: 400px;
-@prompt-form-height: 300px;
-@prompt-form-margin: 24px;
+@import (reference) '../variables';
 
 .content {
   position: absolute;
