+++ static/less/components/prompts/checkbox.less
@@ -1,3 +1,5 @@
+@import (reference) '../../variables';
+
 @prompt-chekbox-item-height: 38px;
 
 .prompt form {
