+++ static/less/components/prompts/confirm.less
@@ -1,3 +1,5 @@
+@import (reference) '../../variables';
+
 .prompt form {
 
   .confirm-options {
