+++ static/less/components/prompts/folder.less
@@ -1,3 +1,5 @@
+@import (reference) '../../variables';
+
 .prompt form {
 
   .select-folder-prompt {
