+++ static/less/components/prompts/input.less
@@ -1,3 +1,5 @@
+@import (reference) '../../variables';
+
 @input-width: 320px;
 
 .prompt form {
