+++ static/less/components/prompts/list.less
@@ -1,3 +1,5 @@
+@import (reference) '../../variables';
+
 @prompt-list-item-height: 16px;
 @prompt-list-item-margin: 38px;
