+++ static/less/variables.less
@@ -1,0 +1,3 @@
+@prompt-form-width: 400px;
+@prompt-form-height: 300px;
+@prompt-form-margin: 24px;
